fix(send): block Review and prompt when the account needs its seed
A sterilized software account (restored backup / missing vault key) can
still build a transaction, then dies at sign with "No stored key for
wallet …". The Accounts page already showed "Needs seed"; Send did not,
so users saw a dead-end confirm error and no recovery affordance.

Detect signability on Send init via getSignableWalletIds /
isAccountSignable. When the current account cannot sign, show a
"Re-enter your recovery phrase" banner, disable Review, and open the
existing import-seed modal in place. Extract that modal so Accounts and
Send share one implementation.

Behavioral tests: sterilized account shows the banner and blocks Review;
a signable account does not.
